Readjusted routes
