+++ controllers/jobsController.js
@@ -8,6 +8,26 @@ function redirectIfNotAuthenticated(req, res, next, userId) {
     res.status(403).render("errors/403");
   }
 }
+
+exports.redirectToBrowse = function(req, res, next) {
+  res.redirect(req.baseUrl + "/browse");
+};
+
+exports.browse = async function(req, res, next) {
+  var user_skills = ["PHP", "Python", "MySQL", "Linux", "JavaScript"];
+  var user_lang = ["Thai", "English"];
+  const jobs = await Job.query();
+  // var jobs = [{id:1,job:'test',job_info:'Lorem',job_type:'Hourly',tag:'Python',price:500},{id:2,job:'test2',job_info:'Lorem2',job_type:'Fixed',tag:'PHP',price:7500}]
+  let title = "Projects | JetFree by JainsBret";
+  res.render("jobs/browse", {
+    title: title,
+    jobs: jobs,
+    n_results: jobs.length,
+    skills: user_skills,
+    lang: user_lang
+  });
+};
+
 exports.view = async function(req, res, next) {
   const job = await Job.query()
     .findById(req.params.jobId)
