+++ controllers/profileController.js
@@ -13,3 +13,10 @@ exports.viewProfile = async function(req, res) {
     user: user
   });
 };
+
+exports.redirectToUserProfile = function(req, res) {
+  if (!req.user) {
+    res.redirect("/");
+  }
+  res.redirect(req.baseUrl + "/" + req.user.username);
+};
